refactor: modularize bot commands
