+++ message_services/discord_service.py
@@ -15,6 +15,7 @@ import requests
 import discord
 import re
 import os
+from commands import vip, info, events, interactions, xp
 
 BIRTHDAY_REGEX = re.compile(r'(\d{1,2})(?:\s?(?:d[eo]|\/|\\|\.)\s?|\s?)(\d{1,2}|(?:janeiro|fevereiro|março|abril|maio|junho|julho|agosto|setembro|outubro|novembro|dezembro))(?:\s?(?:d[eo]|\/|\\|\.)\s?|\s?)(\d{4})')
 MONTHS = ['00','janeiro', 'fevereiro', 'março', 'abril', 'maio', 'junho', 'julho', 'agosto', 'setembro', 'outubro', 'novembro', 'dezembro']
@@ -28,6 +29,11 @@ timezone_offset = -3.0  # Pacific Standard Time (UTC−08:00)
 def now() -> datetime: return (datetime.now(timezone(timedelta(hours=timezone_offset)))).replace(tzinfo=None)
 
 
+vip.setup(bot)
+info.setup(bot)
+events.setup(bot)
+interactions.setup(bot)
+xp.setup(bot)
 
 @bot.event
 async def on_ready():
@@ -281,449 +287,6 @@ async def configForVips():
 
 
 
-####################################################################################################################
-# COMANDOS VIP
-####################################################################################################################
-
-
-@bot.tree.command(name=f'vip-mudar_cor', description=f'Muda a cor do cargo VIP do membro')
-async def changeVipColor(ctx: discord.Interaction, cor: str):
-    userVipRoles = [role.id for role in ctx.user.roles if DISCORD_VIP_ROLES_ID.__contains__(role.id)]
-    if userVipRoles.__len__() != 0:
-        if not re.match(r'^#(?:[a-fA-F0-9]{3}){1,2}$', cor):
-            return await ctx.response.send_message(content='''# Cor invalida!\n 
-Você precisa informar uma cor no formato Hex (#000000).
-Você pode procurar por uma cor em https://htmlcolorcodes.com/color-picker/ e testa-la usando o comando "?color #000000"''', ephemeral=False)
-        if not await colorIsAvailable(cor):
-            return await ctx.response.send_message(content='''Cor inválida! você precisa informar uma cor que não seja muito parecida com a cor de algum cargo da staff''', ephemeral=True)
-        corFormatada = int(cor.replace('#','0x'),16)
-        customRole = await addVipRole(ctx)
-        await customRole.edit(color=corFormatada)
-        await ctx.response.send_message(content=f'Cor do cargo VIP alterada para {cor} com sucesso!', ephemeral=False)
-        return saveCustomRole(ctx.guild_id, ctx.user, color=corFormatada)
-    return await ctx.response.send_message(content='Você não é vip! você não pode fazer isso', ephemeral=True)
-
-@bot.tree.command(name=f'vip-mudar_icone', description=f'Muda o icone do cargo VIP do membro')
-async def changeVipIcon(ctx: discord.Interaction, icon: str):
-    userVipRoles = [role.id for role in ctx.user.roles if DISCORD_VIP_ROLES_ID.__contains__(role.id)]
-    if userVipRoles.__len__() != 0:
-        try:
-            customRole = await addVipRole(ctx)
-            if icon.__contains__('<') or icon.__contains__('>') or icon.__contains__(':'):
-                try:
-                    emoji = ctx.guild.get_emoji(int(icon.replace('<','').replace('>','').split(':')[2]))
-                except:
-                    return await ctx.response.send_message(content='''Ícone inválido! apenas emojis do servidor são permitidos''', ephemeral=True)
-                icon = await emoji.read()
-            await customRole.edit(display_icon=icon)
-            await ctx.response.send_message(content='Ícone do cargo VIP alterado com sucesso!', ephemeral=False)
-            return saveCustomRole(ctx.guild_id, ctx.user, iconId=emoji.id)
-        except Exception as e:
-            print(e)
-            return await ctx.response.send_message(content='''Algo deu errado, avise o titio sobre!''', ephemeral=True)
-    return await ctx.response.send_message(content='Você não é vip! você não pode fazer isso', ephemeral=True)
-
-
-
-
-####################################################################################################################
-# COMANDOS DE INFORMAÇÕES DE MEMBROS
-####################################################################################################################
-
-
-@bot.tree.command(name=f'registrar_local', description=f'Registra o seu local')
-async def registerLocal(ctx: discord.Interaction, local: str):
-    mydb = connectToDatabase()
-    availableLocals = getAllLocals(mydb)
-    if stateLetterCodes[local]:
-        await ctx.response.defer()
-        result = includeLocale(mydb, ctx.guild.id, local.upper(), ctx.user, availableLocals)
-        endConnectionWithCommit(mydb)
-        if result:
-            for locale in availableLocals:
-                if locale['locale_abbrev'] == local.upper():
-                    return await ctx.followup.send(content=f'você foi registrado em {locale["locale_name"]}!', ephemeral=False)
-        else:
-            return await ctx.followup.send(content=f'Não foi possível registrar você! você já está registrado em algum local?', ephemeral=True)
-        return
-    
-    
-@bot.tree.command(name=f'furros_na_area', description=f'Lista todos os furries registrados em um local')
-async def listFurries(ctx: discord.Interaction, local: str):
-    mydb = connectToDatabase()
-    availableLocals = getAllLocals(mydb)
-    if stateLetterCodes[local]:
-        await ctx.response.defer()
-        result = getByLocale(mydb,local.upper(), availableLocals)
-        endConnection(mydb)
-        if result:
-            for locale in availableLocals:
-                if locale['locale_abbrev'] == local.upper():
-                    membersResponse = ',\n'.join(member for member in result)
-                    return await ctx.followup.send(content=f'''Aqui estão os furros registrados em {locale["locale_name"]}:```{membersResponse}```''')
-        else:
-            for locale in availableLocals:
-                if locale['locale_abbrev'] == local.upper():
-                    return await ctx.followup.send(content=f'Não há furros registrados em {locale["locale_name"]}... que tal ser o primeiro? :3')
-    else:
-        return await ctx.response.send_message(content='''Local inválido! Você deve informar uma sigla de Estado válido''', ephemeral=True)
-    
-
-@bot.tree.command(name=f'registrar_aniversario', description=f'Registra seu aniversário')
-async def registerBirthday(ctx: discord.Interaction, data: str, mencionavel: Literal["sim", "não"]):
-    try:
-        datetime.strptime(data, "%d/%m/%Y")
-    except ValueError:
-        return await ctx.response.send_message(content='''Data de nascimento inválida! você informou uma data no formato "dd/mm/aaaa"? <:catsip:851024825333186560>''', ephemeral=True)
-    birthdayAsDate = datetime.strptime(data, '%d/%m/%Y').date()
-    mencionavel = True if mencionavel == "sim" else False
-    await ctx.response.defer()
-    mydb = connectToDatabase()
-    try:
-        registered = includeBirthday(mydb, ctx.guild.id, birthdayAsDate, ctx.user, mencionavel, None, True)
-        if registered:
-            return await ctx.followup.send(content=f'você foi registrado com o aniversário {birthdayAsDate.day:02}/{birthdayAsDate.month:02}!', ephemeral=False)
-        else:
-            return await ctx.followup.send(content=f'Algo deu errado... Avise o titio!', ephemeral=False)
-    except Exception as e:
-        if str(e).__contains__('Duplicate entry'):
-            return await ctx.followup.send(content=f'Você ja está registrado. Caso o seu aniversário não esteja aparecendo na lista, tente usar /{ctx.command} com mencionável = Sim', ephemeral=True)
-        if str(e).__contains__('Changed Entry'):
-            if mencionavel:
-                return await ctx.followup.send(content=f'Seu aniversário foi atualizado para ser mencionável!', ephemeral=False)
-            else:
-                return await ctx.followup.send(content=f'Seu aniversário foi atualizado para não ser mencionável!', ephemeral=False)
-        return await ctx.followup.send(content=f'Algo deu errado... Avise o titio!', ephemeral=False)
-    finally:
-        endConnectionWithCommit(mydb)
-    
-
-@bot.tree.command(name=f'aniversarios', description=f'Lista todos os aniversários registrados')
-async def listBirthdays(ctx: discord.Interaction):
-    await ctx.response.defer()
-    result = getAllBirthdays()
-    if result:
-        # troca os userIDs por usernames
-        for birthday in result:
-            birthday['user'] = ctx.guild.get_member(birthday['user_id'])
-        birthdaysResponse = ',\n'.join(
-    f'{birthday["birth_date"].strftime("%d/%m")} - {birthday["user"].display_name}'
-    for birthday in sorted(result, key=lambda birthday: (birthday["birth_date"].month, birthday["birth_date"].day)) if birthday["user"] != None)
-        return await ctx.followup.send(content=f'Aqui estão os aniversários registrados:```{birthdaysResponse}```')
-    else:
-        return await ctx.followup.send(content=f'Não há aniversários registrados... que tal ser o primeiro? :3')
-
-
-####################################################################################################################
-# COMANDOS DE EVENTOS
-####################################################################################################################
-
-
-async def addEvent(ctx: discord.Interaction, user: any, state:str, city:str, event_name:str, address:str, price:float, starting_date: str, starting_time: str, ending_date: str, ending_time: str, description: str = None, group_link:str = None, site:str = None, max_price:float = None, event_logo_url:str = None):
-    try:
-        datetime.strptime(starting_date, "%d/%m/%Y")
-        if ending_date == None:
-            ending_date = starting_date
-        else:
-            datetime.strptime(ending_date, "%d/%m/%Y")
-    except ValueError:
-        return await ctx.response.send_message(content='''Data inválida! você informou uma data no formato "dd/mm/aaaa"? <:catsip:851024825333186560>''', ephemeral=True)
-    try:
-        datetime.strptime(starting_time, "%H:%M")
-        datetime.strptime(ending_time, "%H:%M")
-    except ValueError:
-        return await ctx.response.send_message(content='''Horário inválido! você informou um horário no formato "hh:mm"? <:catsip:851024825333186560>''', ephemeral=True)
-    starting_datetime = datetime.strptime(f'{starting_date} {starting_time}', '%d/%m/%Y %H:%M')
-
-    ending_datetime = datetime.strptime(f'{ending_date} {ending_time}', '%d/%m/%Y %H:%M')
-    if starting_datetime > ending_datetime:
-        #teste se a data e hora de inicio é maior que a data e hora de encerramento
-        return await ctx.response.send_message(content='''A data e hora de início do evento não pode ser maior que a data e hora de encerramento!''', ephemeral=True)
-    mydb = connectToDatabase()
-    locale_id = await localeIsAvailable(ctx, mydb, state)
-    if locale_id:
-        await ctx.response.defer()
-        result = includeEvent(mydb,user, locale_id, city, event_name, address, price, starting_datetime, ending_datetime, description, group_link, site, max_price, event_logo_url)
-        endConnectionWithCommit(mydb)
-        if result:
-            return await ctx.followup.send(content=f'O evento **{event_name}** foi registrado com sucesso!', ephemeral=False)
-        
-@bot.tree.command(name=f'novo_evento', description=f'Adiciona um evento ao calendário')
-async def addEventWithDiscordUser(ctx: discord.Interaction, user: discord.Member, estado:str, cidade:str, event_name:str, address:str, price:float, starting_date: str, starting_time: str, ending_date: str, ending_time: str, description: str = None, group_link:str = None, site:str = None, max_price:float = None, event_logo_url:str = None):
-    await addEvent(ctx, user, estado, cidade, event_name, address, price, starting_date, starting_time, ending_date, ending_time, description, group_link, site, max_price, event_logo_url)
-    pass
-@bot.tree.command(name=f'novo_evento_por_usuario', description=f'Adiciona um evento ao calendário usando um usuário do telegram')
-async def addEventWithTelegramUser(ctx: discord.Interaction, telegram_username: str, estado:str, cidade:str, event_name:str, address:str, price:float, starting_date: str, starting_time: str, ending_time: str, ending_date: str = None, description: str = None, group_link:str = None, site:str = None, max_price:float = None, event_logo_url:str = None):
-    await addEvent(ctx, telegram_username, estado, cidade, event_name, address, price, starting_date, starting_time, ending_date, ending_time, description, group_link, site, max_price, event_logo_url)
-    pass
-
-
-@bot.tree.command(name=f'eventos', description=f'Lista todos os eventos registrados')
-async def listEvents(ctx: discord.Interaction):
-    await ctx.response.defer()
-    result = getAllEvents()
-    if result:
-        formattedEvents = formatEventList(result)
-        eventsResponse = []
-        for event in formattedEvents:
-            if eventsResponse == []:
-                eventsResponse.append(f'''Aqui estão os próximos eventos registrados:\n'''+event+'\n‎')
-            else:
-                eventsResponse.append('\n\n'+ event)
-                if event != formattedEvents[-1]:
-                    eventsResponse[-1] += '\n‎'
-        eventsResponse[-1] += f'''\n\n```Se você quiser ver mais detalhes sobre um evento, use o comando "/evento <nome do evento>"```
-Adicione tambem a nossa agenda de eventos ao seu google agenda e tenha todos os eventos na palma da sua mão! {os.getenv('GOOGLE_CALENDAR_LINK')}'''
-        for message in eventsResponse:
-            await ctx.followup.send(content=message) if message != eventsResponse[-1] else await ctx.channel.send(content=message)
-    else:
-        return await ctx.followup.send(content=f'Não há eventos registrados... que tal ser o primeiro? :3')
-    
-    
-@bot.tree.command(name=f'eventos_por_estado', description=f'Lista todos os eventos registrados em um estado')
-async def listEventsByState(ctx: discord.Interaction, state: str):
-    mydb = connectToDatabase()
-    await ctx.response.defer()
-    locale_id = await localeIsAvailable(ctx, mydb, state)
-    result = getEventsByState(locale_id)
-    endConnection(mydb)
-    if result:
-        formattedEvents = formatEventList(result)
-        eventsResponse = []
-        for event in formattedEvents:
-            if eventsResponse == []:
-                eventsResponse.append(f'''Aqui estão os próximos eventos registrados:\n'''+event+'\n‎')
-            else:
-                eventsResponse.append('\n\n'+ event)
-                if event != formattedEvents[-1]:
-                    eventsResponse[-1] += '\n‎'
-        eventsResponse[-1] += f'''\n\n```Se você quiser ver mais detalhes sobre um evento, use o comando "/evento <nome do evento>"```
-Adicione tambem a nossa agenda de eventos ao seu google agenda e tenha todos os eventos na palma da sua mão! {os.getenv('GOOGLE_CALENDAR_LINK')}'''
-        for message in eventsResponse:
-            await ctx.followup.send(content=message) if message != eventsResponse[-1] else await ctx.channel.send(content=message)
-    else:
-        return await ctx.followup.send(content=f'Não há eventos registrados em {state}... que tal ser o primeiro? :3')
-    
-    
-@bot.tree.command(name=f'evento', description=f'Mostra os detalhes de um evento')
-async def showEvent(ctx: discord.Interaction, event_name: str):
-    if event_name.__len__() < 4:
-        return await ctx.response.send_message(content='''Nome do evento inválido! você informou um nome com menos de 4 caracteres? <:catsip:851024825333186560>''', ephemeral=True)
-    await ctx.response.defer()
-    event = getEventByName(event_name)
-    if event:
-        eventEmbeded = formatSingleEvent(event)
-        return await ctx.followup.send(embed=eventEmbeded)
-    else:
-        return await ctx.followup.send(content=f'Não há eventos registrados com esse nome. Tem certeza que digitou o nome certo?')
-    
-
-@bot.tree.command(name=f'evento_reagendar', description=f'Reagenda um evento pendente')
-async def rescheduleEvent(ctx: discord.Interaction, event_name: str, new_date: str):
-    try: new_date = datetime.strptime(new_date, "%d/%m/%Y")
-    except ValueError:
-        return await ctx.response.send_message(content='''Data inválida! você informou uma data no formato "dd/mm/aaaa"? <:catsip:851024825333186560>''', ephemeral=True)
-    await ctx.response.defer()
-    result = rescheduleEventDate(event_name, new_date, ctx.user.name)
-    if result == True:
-        return await ctx.followup.send(content=f'O evento **{event_name}** foi reagendado com sucesso!')
-    elif result == "não encontrado":
-        return await ctx.followup.send(content=f'Não foi possível encontrar um evento com esse nome! tem certeza que digitou o nome certo?')
-    elif result == "não é o dono":
-        return await ctx.followup.send(content=f'Você não é o dono desse evento! apenas o dono pode reagendar o evento')
-    elif result == "em andamento":
-        return await ctx.followup.send(content=f'Não é possível reagendar um evento que já está em andamento')
-    elif result == "encerrado":
-        return await ctx.followup.send(content=f'Não é possível reagendar um evento que já foi realizado')
-
-
-@bot.tree.command(name=f'evento_agendar_prox', description=f'Agenda no calendario a próxima data do evento')
-async def scheduleNextEvent(ctx: discord.Interaction, nome_do_evento: str, data: str):
-    try: data = datetime.strptime(data, "%d/%m/%Y")
-    except ValueError:
-        return await ctx.response.send_message(content='''Data inválida! Você informou uma data no formato "dd/mm/aaaa"? <:catsip:851024825333186560>''', ephemeral=True)
-    await ctx.response.defer()
-    result = scheduleNextEventDate(nome_do_evento, data, ctx.user.name)
-    if result == True:
-        return await ctx.followup.send(content=f'O evento **{nome_do_evento}** foi agendado com sucesso!')
-    elif result == "não encontrado":
-        return await ctx.followup.send(content=f'Não foi possível encontrar um evento com esse nome! tem certeza que digitou o nome certo?')
-    elif result == "não encerrado":
-        return await ctx.followup.send(content=f'Não é possível agendar um evento que ainda não foi encerrado')
-    elif result == "não é o dono":
-        return await ctx.followup.send(content=f'Você não é o dono desse evento! apenas o dono pode agendar o evento')
-    
-
-@bot.tree.command(name=f'eventos_pendentes', description=f'Mostra os eventos esperando aprovação')
-async def showPendingEvents(ctx: discord.Interaction):
-    await ctx.response.defer()
-    result = getAllPendingApprovalEvents()
-    if result:
-        eventsResponse = '\n\n'.join(
-    f'''> # {event["event_name"].title()} (id - {event["id"]})
->    **Data**: {event["starting_datetime"].strftime("%d/%m/%Y") if event["starting_datetime"].strftime("%d/%m/%Y") == event["ending_datetime"].strftime("%d/%m/%Y") else f"{event['starting_datetime'].strftime('%d')} a {event['ending_datetime'].strftime('%d/%m/%Y')}"}{" - das "+event["starting_datetime"].strftime("%H:%M")+" às "+event["ending_datetime"].strftime("%H:%M") if event["starting_datetime"] == event["ending_datetime"] else ''}
->    **Local**: {event["city"]}, {event["state_abbrev"]}
->    **Endereço**: {event["address"]} ''' + '\n'+
-    '\n'.join(filter(None, [
-        f">    **Chat do evento**: <{event['group_chat_link']}>" if event['group_chat_link']!=None else '',
-        f">    **Site**: <{event['website']}>" if event['website']!=None else '',
-        f'''>    **Preço**: {"A partir de R$"+str(f"{event['price']:.2f}").replace('.',',') if event['price']!=0 else 'Gratuito'}'''
-        ]))
-        for event in sorted(result, key=lambda event: event["starting_datetime"]))
-        await ctx.followup.send(content=f'''Aqui estão os eventos registrados esperando aprovação:\n{eventsResponse}\n
-```Se você quiser aprovar um evento, use o comando "/evento_aprovar <id do evento>"```''')
-    else:
-        return await ctx.followup.send(content=f'Não há eventos a serem aprovados... talvez seja a hora de buscar novos eventos!')
-
-
-@bot.tree.command(name=f'evento_aprovar', description=f'Aprova um evento pendente')
-async def approveEvent(ctx: discord.Interaction, event_id: int):
-    if ctx.user.id != 167436511787220992:
-        return await ctx.response.send_message(content='Você não tem permissão para fazer isso', ephemeral=True)
-    await ctx.response.defer()
-    result = approveEventById(event_id)
-    if result == True:
-        return await ctx.followup.send(content=f'O evento foi aprovado com sucesso!')
-    elif result == "não encontrado":
-        return await ctx.followup.send(content=f'Não foi possível encontrar um evento pendente com esse id! tem certeza que digitou o id certo?')
-    elif result == False:
-        return await ctx.followup.send(content=f'Não foi possível aprovar o evento')
-
-
-@bot.tree.command(name=f'evento_add_staff', description=f'Adiciona um membro da staff como organizador de um evento')
-async def addStaffToEvent(ctx: discord.Interaction, event_id: int, staff: discord.Member):
-    pass
-
-
-####################################################################################################################
-# COMANDO DE CONEXÃO DE CONTAS
-####################################################################################################################
-
-
-@bot.tree.command(name=f'adm-conectar_conta', description=f'Conecta sua conta do discord com a do telegram')
-async def connectAccount(ctx: discord.Interaction,user: discord.Member, telegram_username: str):
-    await ctx.response.defer()
-    result = admConnectTelegramAccount(ctx.guild.id, user, telegram_username)
-    if result:
-        return await ctx.followup.send(content=f'Sua conta foi conectada com sucesso! agora você pode usar os comandos do bot no discord e no telegram', ephemeral=False)
-    else:
-        return await ctx.followup.send(content=f'Não foi possível conectar sua conta! você já está conectado?', ephemeral=True)
-
-
-####################################################################################################################
-# COMANDOS DE INTERAÇÕES DO CODDY
-####################################################################################################################
-
-
-@bot.tree.command(name=f'{BOT_NAME.lower()}_diz', description=f'Faz {BOT_NAME} falar em um canal de texto')
-async def sayAsCoddy(ctx: discord.Interaction, channel: discord.TextChannel, message: str):
-    channelId = discord.utils.get(ctx.guild.channels, name=channel.name)
-    await channelId.send(message)
-    resp = await ctx.response.send_message(content='mensagem enviada!', ephemeral=True)
-    return resp
-
-@bot.tree.command(name=f'{BOT_NAME.lower()}_status', description=f'Muda o status do {BOT_NAME}')
-async def changeMood(ctx: discord.Interaction, mood: Literal['jogando','ouvindo','assistindo','mensagem'], message: str):
-    moodDiscord = 'playing' if mood == 'jogando' else 'listening' if mood == 'ouvindo' else 'watching'
-    if mood == 'mensagem':
-        await bot.change_presence(activity=discord.CustomActivity(name=message))
-        resp = await ctx.response.send_message(content=f'{BOT_NAME} está com o status de "{message}"', ephemeral=True)
-    else:
-        await bot.change_presence(activity=discord.Activity(type=getattr(discord.ActivityType, moodDiscord), name=message))
-        resp = await ctx.response.send_message(content=f'{BOT_NAME} está {mood} {message}!', ephemeral=True)
-    return resp
-
-
-####################################################################################################################
-# COMANDOS DE XP
-####################################################################################################################
-
-
-@bot.tree.command(name=f'xp', description=f'Mostra a quantidade de xp de um membro')
-async def showXp(ctx: discord.Interaction, member: discord.Member):
-    pass
-
-@bot.tree.command(name=f'xp_ranking', description=f'Mostra o ranking de xp dos membros')
-async def showXpRanking(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'xp_resetar', description=f'Resetar a quantidade de xp de um membro')
-async def resetXp(ctx: discord.Interaction, member: discord.Member):
-    pass
-
-@bot.tree.command(name=f'xp_resetar_todos', description=f'Resetar a quantidade de xp de todos os membros')
-async def resetAllXp(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'xp_adicionar', description=f'Adiciona xp a um membro')
-async def addXp(ctx: discord.Interaction, member: discord.Member, xp: int):
-    pass
-
-@bot.tree.command(name=f'xp_remover', description=f'Remove xp de um membro')
-async def removeXp(ctx: discord.Interaction, member: discord.Member, xp: int):
-    pass
-
-###############################################################
-
-@bot.tree.command(name=f'loja', description=f'Compre itens com seu dinheiro')
-async def shop(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'rp_inventario', description=f'Mostra os itens que você possui')
-async def inventory(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'rp_usar', description=f'Usa um item do seu inventário')
-async def useItem(ctx: discord.Interaction, item: str):
-    pass
-
-@bot.tree.command(name=f'rp_vender', description=f'Vende um item do seu inventário')
-async def sellItem(ctx: discord.Interaction, item: str):
-    pass
-
-################################################################
-
-@discord.app_commands.checks.cooldown(1, 86400, key=lambda ctx: (ctx.guild_id, ctx.author.id))
-@bot.tree.command(name=f'daily', description=f'Pega sua recompensa diária')
-async def daily(ctx: discord.Interaction):
-    await ctx.response.send_message(content='Recompensa diária pega com sucesso!')
-    pass
-
-@daily.error
-async def daily_error(ctx, error):
-    if isinstance(error, commands.CommandOnCooldown):
-        await ctx.send(f'Você já pegou sua recompensa diária! Tente novamente em {error.retry_after:.0f} segundos', ephemeral=True)
-    else:
-        raise error
-
-@bot.tree.command(name=f'rp_banho', description=f'Tomar banho dá xp sabia?')
-async def bath(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'rp_trabalhar', description=f'Trabalha em troca de xp e dinheiro')
-async def work(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'rp_duelo', description=f'Desafie alguém para um duelo')
-async def duel(ctx: discord.Interaction, member: discord.Member):
-    pass
-
-@bot.tree.command(name=f'rp_desenhar', description=f'Desenhe algo!')
-async def draw(ctx: discord.Interaction):
-    pass
-
-@bot.tree.command(name=f'rp_escrever', description=f'Escreva uma história')
-async def write(ctx: discord.Interaction):
-    pass
-
-""" @bot.tree.command(name=f'rp_missao', description=f'Complete missões para ganhar xp e dinheiro')
-async def mission(ctx: discord.Interaction):
-    pass """
-
-
-
-####################################################################################################################
-# COMANDOS DE PERFIL
 ####################################################################################################################
